Add unit test SSLConnectTimeout
Add unit test for issue #682 fixed in PR#728, which does not contain
the test of its own.

The test creates a fake SSL server which does not create an SSL context.
When an SSL client attempts to send it a request, it gets a timeout
error. Prior to PR#682, the client would wait indefinitely.