Add the capability to decommission a Server object
Fixes #1645.

`Server::stop` does nothing if the server is not running.
This can cause a program to hang it it calls stop before
the server processing thread has had time to initialize.
It would make sense for a server that has been stopped
to refuse to start, but that would be backwards incompatible.
Instead, we introduce a new method: `decommission`.
It sets the server object to stop and refuse to start.
This means it does not matter in which order
concurrent threads end up calling `listen` and `decommission`,
the result is always that the listening thread will exit.