Add client stream handler
Introduce a new request callback, stream_handler. If set, it replaces
all other content-serving mechanisms. The handler is not called in
response to HEAD or CONNECT requests, or when following a redirect.
Content-related default header fields are not added to the request.

In conjunction with a response handler, it provides a minimal interface
to implement a WebSocket client.